[Bronze V] Title: 이상한 기호, Time: 132 ms, Memory: 14196 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,7 +8,7 @@ public class Main {
     static StringBuffer sb = new StringBuffer();
     static StringTokenizer st;
 
-    static int a, b;
+    static long a, b;
 
     static void input() throws IOException{
         st = new StringTokenizer(br.readLine());
